Added new Recommender -- NMTopRecommender
Selects the best N results for the closest M files. The best agent is
selected as the most often used agent on those M files.

The average and stardard deviation of individual option values is computed and the recommended
agent uses the range <avg - 2*sd,avg + 2*sd> for each of these options.

Options with sd = 0 use the value of avg (no range, immutable).